tools/jaklis/requirements.txt
wheel
base58
pybase64
duniterpy==0.62.0
termcolor
python-dotenv
gql==3.0.0a5
#gql==2.0
requests
